backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
PyJWT
requests
psycopg2-binary
gunicorn
schedule